Makes docker image use snapshot instead of evergreen repository.
The `sed` just comments out the evergreen package repository and uncomments the snapshot repository associated with the base image.

The snapshot repositories automatically "expire", so we have to tell apt to ignore expiration dates with `-o Acquire::Check-Valid-Until=false`.

Rolled back curl by one version so we are using the version that was available at the time the base debian image was built.